Create the TaskTag ORM class, re-organize the reflection code
