+++ src/database/schema/__init__.py
@@ -1,0 +1,1 @@
+"""Defines Object-Relational Mappings (ORM)."""
